THE ROAD APP/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
GeoAlchemy2==0.14.2
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0